Deprecate WebConfiguration::getDefaultsForWiki()
I've spent an hour or so reading the code, and testing it, and I have seen no evidence to indicate that this actually gets the defaults for a wiki, rather than what it appears to do, which is getting the *overridden* settings for a foreign wiki.

This is causing bugs on at least one of my installs.